Make nes-emu build and run on Linux (gcc and clang)
- Fixed compilation errors
- Added build_linux.sh that invokes CMake to generate gcc and clang makefiles and builds both

+++ src/main.cpp
@@ -33,9 +33,9 @@ namespace
 	{
 		printf("Rom Info:\n");
 		printf("  File: %s\n", romFile);
-		printf("  PRG ROM size: %d kb\n", BytesToKB(header.GetPrgRomSizeBytes()));
-		printf("  CHR ROM size: %d kb\n", BytesToKB(header.GetChrRomSizeBytes()));
-		printf("  PRG RAM size: %d kb\n", BytesToKB(header.GetPrgRamSizeBytes()));
+		printf("  PRG ROM size: %d kb\n", (int)BytesToKB(header.GetPrgRomSizeBytes()));
+		printf("  CHR ROM size: %d kb\n", (int)BytesToKB(header.GetChrRomSizeBytes()));
+		printf("  PRG RAM size: %d kb\n", (int)BytesToKB(header.GetPrgRamSizeBytes()));
 		printf("  Mapper number: %d\n", header.GetMapperNumber());
 		printf("  Has SRAM: %s\n", header.HasSRAM()? "yes" : "no");
 		printf("\n");
